Improve action button behavior
If we click the Diff cell and move mouse to another cell, while still holding down the left mouse button, then we no longer treat that other cell as a hit target.

This commit minimizes number of changes to achieve the behavior. The entire handling of the event loop can be simplified, partly based on the changed behavior.